requirements.txt
numpy>=1.15.0, <1.21.0
pydantic == 1.10.4
pytest == 7.2.1
torch == 1.13.1
torchvision == 0.14.1
nltk == 3.8.1
pyyaml == 6.0
